PYTHON/waveshare/requirements.txt
# UC2 CANopen Python Tool — Requirements
#
# Install with:
#   pip install -r requirements.txt

pyserial>=3.5
python-can>=4.3.0
canopen>=2.2.0
